# MAINT: Improve speed of ufunc kwargs parsing.