… numpy/core/src/umath/ufunc_object.c