This uses the realization from inspecting similar CPython code that as we intern string names anyway, comparing keys with possible names by pointer will generally just work, and is of course much faster than first converting a unicode object to `char*` and then using `strcmp`.

@eric-wieser - my C remains limited: is there a cleverer way than the `enum` I used, ideally one that allows on to type things just once (there are now five places one has to keep up to date, three for the interns, one enum, and one array of the interned strings; this seems too much!)

Reduces the overhead as follows:
```
import numpy as np
a = np.array(1)
b = np.empty_like(a)
%timeit np.positive(a)                                 # 352->348 ns
%timeit np.positive(a, subok=True)                     # 606->501 ns
%timeit np.positive(a, where=True)                     # 586->503 ns
%timeit np.positive(a, where=True, subok=True)         # 695->531 ns
%timeit np.positive(a, b)                              # 354->352 ns
%timeit np.positive(a, out=b)                          # 557->480 ns
%timeit np.positive(a, out=b, subok=True)              # 668->506 ns
%timeit np.positive(a, out=b, where=True, subok=True)  # 752->536 ns
```