@eric-wieser 2018-06-15
I'm curious how much the `switch` statement buys us over a bunch of `if`/`else if`s comparing `PyObject *` pointers to interned strings, which would eliminate the need for an enum.