@mhvk 2018-06-15
The one issue would be that then the slow path becomes a whole set of `goto` statements... (and given that python has it, I do not quite dare to remove it...)