@mhvk 2018-06-15
I think the real solution here would be to mimic `PyArg_ParseTupleAndKeywords`, with all those bits of code interpreting the various keywords properly passed on as functions. But that may be better done as another PR...